Remove SPicklerUnpickler type entirely
